sphinx_reload: deconflict watch and documentation_root args
When users declare a `--watch` argument at the end of their argv string,
argparse will assume that *all* words thereafter are watch directories.
This is because the `--watch` argument is set to `nargs=*`. As a result,
the `documentation_root` positional argument is erroneously consumed as
a watch element and sphinx_reload immediately exits 2, repoting a
missing argument.

Instead, use the default `nargs` value of 1, and set the argument action
to "append". Now, users will have to specify `--watch` before *each*
directory sting that they want to watch, but at least the arguments
work.

Signed-off-by: Alex Stewart <astewart.49c6@gmail.com>